feat: Start work on the tool
This includes the first steps in creating the tool.

It now has a CLI-parser, generated using `docopts`. This allows keeping the
definition of the CLI in a single file and makes for a more config-driven
approach, which I find pretty neat!

The rest is some scaffolding around that - mainly that complicated meson-setup
for codegen. This may be able to be improved.. I don't know. It's kinda ugly.