feat: Add support for parsing unencrypted keystore-contents
We follow the recipe laid out by the format-specification and pull out
the fields from the keystore.

The keyring decode and marshal functionality is contained in a new lib.
This allows testing the keyring-module individually, for example.

Everything that has can be unmarshalled so far can also be dumped in
JSON-format (and be processed further using, for example `jq`)! :)